requirements.txt
# Core Framework
Flask==3.1.1
Flask-RESTful==0.3.10
flask-cors==6.0.0
flask-request-id-header==0.1.1
Werkzeug==3.1.3

# Discord Bot
discord.py==2.5.2

# Web and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.11.14
httpx==0.28.1

# Reddit Integration
praw==7.8.1
prawcore==2.4.0

# AI/ML and Language Processing
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
google-generativeai==0.8.4
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
duckduckgo_search==8.0.2

# Data Processing and Validation
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1

# Async Support
asyncio==3.4.3
nest-asyncio==1.6.0

# Utilities
python-dateutil==2.9.0.post0
tqdm==4.67.1
rich==14.0.0